# server/api/openapi.gen.yaml
...
components:
...
    responses:
...
        Forbidden:
            content:
                application/json:
                    schema:
                        $ref: '#/components/schemas/Error'
            description: Authenticated but not permitted
...
        Unauthorized:
            content:
                application/json:
                    schema:
                        $ref: '#/components/schemas/Error'
            description: Not authenticated
...
    schemas:
...
        CatalogStatus:
            enum:
                - active
                - archived
            type: string
...
        CreatePartnerRequest:
            properties:
                code:
                    type: string
                isCustomer:
                    default: false
                    type: boolean
                isSupplier:
                    default: false
                    type: boolean
                name:
                    minLength: 1
                    type: string
            required:
                - name
            type: object
...
        Partner:
            properties:
                code:
                    description: Optional; unique within a tenant when present
                    type: string
                id:
                    format: uuid
                    type: string
                isCustomer:
                    type: boolean
                isSupplier:
                    type: boolean
                name:
                    type: string
                status:
                    $ref: '#/components/schemas/CatalogStatus'
            required:
                - id
                - name
                - isSupplier
                - isCustomer
                - status
            type: object
...
    securitySchemes:
        sessionCookie:
            in: cookie
            name: sinta_session
            type: apiKey
info:
    description: |
        Multi-tenant inventory management for Indonesian SME distributors. Contract-first: this file is the single source of truth for both the Go server (oapi-codegen) and the TypeScript client (openapi-typescript). Conventions: camelCase fields, kebab-case plural paths, /v1 prefix (docs/reference/CONVENTIONS.md).
    title: Sinta API
    version: 0.1.0
openapi: 3.0.3
paths:
...
    /partners:
        get:
            operationId: listPartners
            parameters:
                - description: Filter to partners that are a supplier or a customer
                  in: query
                  name: role
                  required: false
                  schema:
                    enum:
                        - supplier
                        - customer
                    type: string
                - in: query
                  name: status
                  required: false
                  schema:
                    $ref: '#/components/schemas/CatalogStatus'
                - in: query
                  name: q
                  required: false
                  schema:
                    type: string
            responses:
                "200":
                    content:
                        application/json:
                            schema:
                                items:
                                    $ref: '#/components/schemas/Partner'
                                type: array
                    description: Partners
                "401":
                    $ref: '#/components/responses/Unauthorized'
                "403":
                    $ref: '#/components/responses/Forbidden'
            summary: List partners in the active tenant
        post:
            operationId: createPartner
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/CreatePartnerRequest'
                required: true
            responses:
                "201":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Partner'
                    description: Partner created
                "401":
                    $ref: '#/components/responses/Unauthorized'
                "403":
                    $ref: '#/components/responses/Forbidden'
                "409":
                    $ref: '#/components/responses/Conflict'
                "422":
                    $ref: '#/components/responses/ValidationError'
            summary: Create a partner
    /partners/{partnerId}:
        get:
            operationId: getPartner
            responses:
                "200":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Partner'
                    description: Partner
                "401":
                    $ref: '#/components/responses/Unauthorized'
                "403":
                    $ref: '#/components/responses/Forbidden'
                "404":
                    $ref: '#/components/responses/NotFound'
            summary: Get a partner
        parameters:
            - in: path
              name: partnerId
              required: true
              schema:
                format: uuid
                type: string
        patch:
            operationId: updatePartner
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/UpdatePartnerRequest'
                required: true
            responses:
                "200":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Partner'
                    description: Updated partner
                "401":
                    $ref: '#/components/responses/Unauthorized'
                "403":
                    $ref: '#/components/responses/Forbidden'
                "404":
                    $ref: '#/components/responses/NotFound'
                "409":
                    $ref: '#/components/responses/Conflict'
                "422":
                    $ref: '#/components/responses/ValidationError'
            summary: Update a partner
    /products:
        get:
            operationId: listProducts
            parameters:
                - in: query
                  name: status
                  required: false
                  schema:
                    $ref: '#/components/schemas/CatalogStatus'
                - in: query
                  name: q
                  required: false
                  schema:
                    type: string
            responses:
                "200":
                    content:
                        application/json:
                            schema:
                                items:
                                    $ref: '#/components/schemas/Product'
                                type: array
                    description: Products
                "401":
                    $ref: '#/components/responses/Unauthorized'
                "403":
                    $ref: '#/components/responses/Forbidden'
            summary: List products in the active tenant
        post:
            operationId: createProduct
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/CreateProductRequest'
                required: true
            responses:
                "201":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Product'
                    description: Product created
                "401":
                    $ref: '#/components/responses/Unauthorized'
                "403":
                    $ref: '#/components/responses/Forbidden'
                "409":
                    $ref: '#/components/responses/Conflict'
                "422":
                    $ref: '#/components/responses/ValidationError'
            summary: Create a product
    /products/{productId}:
        get:
            operationId: getProduct
            responses:
                "200":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Product'
                    description: Product
                "401":
                    $ref: '#/components/responses/Unauthorized'
                "403":
                    $ref: '#/components/responses/Forbidden'
                "404":
                    $ref: '#/components/responses/NotFound'
            summary: Get a product
        parameters:
            - $ref: '#/components/parameters/ProductId'
        patch:
            operationId: updateProduct
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/UpdateProductRequest'
                required: true
            responses:
                "200":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Product'
                    description: Updated product
                "401":
                    $ref: '#/components/responses/Unauthorized'
                "403":
                    $ref: '#/components/responses/Forbidden'
                "404":
                    $ref: '#/components/responses/NotFound'
                "409":
                    $ref: '#/components/responses/Conflict'
                "422":
                    $ref: '#/components/responses/ValidationError'
            summary: Update a product
    /products/{productId}/batches:
        get:
            operationId: listProductBatches
            responses:
                "200":
                    content:
                        application/json:
                            schema:
                                items:
                                    $ref: '#/components/schemas/Batch'
                                type: array
                    description: Batches
                "401":
                    $ref: '#/components/responses/Unauthorized'
                "403":
                    $ref: '#/components/responses/Forbidden'
                "404":
                    $ref: '#/components/responses/NotFound'
            summary: List a product's batches
        parameters:
            - $ref: '#/components/parameters/ProductId'
        post:
            operationId: createBatch
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/CreateBatchRequest'
                required: true
            responses:
                "201":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Batch'
                    description: Batch created
                "401":
                    $ref: '#/components/responses/Unauthorized'
                "403":
                    $ref: '#/components/responses/Forbidden'
                "404":
                    $ref: '#/components/responses/NotFound'
                "409":
                    $ref: '#/components/responses/Conflict'
                "422":
                    $ref: '#/components/responses/ValidationError'
            summary: Create a batch for a batch-tracked product
...
    /products/{productId}/uoms/{uomId}:
        delete:
            operationId: deleteProductUom
            responses:
                "204":
                    description: Unit conversion removed
                "401":
                    $ref: '#/components/responses/Unauthorized'
                "403":
                    $ref: '#/components/responses/Forbidden'
                "404":
                    $ref: '#/components/responses/NotFound'
            summary: Remove a unit conversion
        parameters:
            - $ref: '#/components/parameters/ProductId'
            - in: path
              name: uomId
              required: true
              schema:
                format: uuid
                type: string
...
security:
    - sessionCookie: []
servers:
    - url: /v1